requirements.txt
python-dotenv
streamlit  
numpy  
pandas  
matplotlib  
scipy  
requests  
Pillow  
azure-core  
azure-search-documents
azure-identity
azure-storage-blob
azure-storage-queue
azure-storage-file-datalake
azure-storage-file-share
psycopg2-binary
pydantic
